refactor: static 폴더 constants로 수정
